Eliane
Feedback, slow beatings, highs and lows

This drone is seven sines phase-modulating each other in a ring.

The sines are based around a natural minor scale (ET), jumping by a fifth
or a diminished fifth each to the next. And each sine is phase modulated
by the next sine, and the one after that, until the highest sines are phase
modulated by the lowest.