Grove
There is no conductor.

This is a drone inspired by the Mannequins Mangrove, and a conversation with
infinitedigits in which we together worked out how to make a formant oscillator
like that in supercollider. As "curve" is also modulated, it also has some
similarity to the sound of a Just Friends.

Like a real mangrove, as the formant length is modulated, the sound skips between
fundamental frequencies in the utone series.

Unlike a real Mangrove, each formant oscillator is triggered by the zero crossings
of two mixed sines, rather than a single pulse train, which gives unstable-sounding i
patterns of pulses sometimes.

There are five of these oscillators, though they don't usually all play at once.

The envelopes on the five oscillators follow the same principles as the pulse
trains themselves. One addition: the envelopes have a flip-flop between their two
frequencies, which gives them a euclidean rhythm nature.